+++ services/planGestion/src/main/java/com/pfc/planGestion/repositorio/persistencia/jpa/MovimientoJpaRepository.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.pfc.planGestion.repositorio.persistencia.component.MovimientoEntity;
+import com.pfc.planGestion.repositorio.persistencia.entity.MovimientoEntity;
 
-public interface MovimientoJpaRepository extends JpaRepository<MovimientoEntity, Integer>  {
+public interface MovimientoJpaRepository extends JpaRepository<MovimientoEntity, Long>  {
 	List<MovimientoEntity> findByIban_Iban(String iban);
 
 }
